+++ src-tauri/src/helpers/sys_tray.rs
@@ -35,7 +35,6 @@ pub fn generate_system_tray(app: &tauri::AppHandle) -> Result<TrayIcon, tauri::E
         .build()?;
     TrayIconBuilder::with_id("main_tray")
         .tooltip(t("tray.tooltip"))
-        .icon(app.default_window_icon().unwrap().clone())
         .menu(&menu)
         .show_menu_on_left_click(true)
         .on_menu_event(|app_handle, event| match event.id.as_ref() {