+++ CHANGELOG.md
@@ -19,6 +19,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Improved
 - Credits Section
 - System Tray Icon based on a real-time scan setting
+### Fixed
+- The GUI Updater
+### Changed
+- `console.error()` into `toast.error()` across every app
 
 ## [1.0.0-3] - 2026-02-10 (v1 Beta #3)
 ### Translated